+++ src/zenml/artifacts/utils.py
@@ -31,13 +31,7 @@ from typing import (
 )
 from uuid import UUID, uuid4
 
-from zenml.artifacts.preexisting_data_materializer import (
-    PreexistingDataMaterializer,
-)
-from zenml.client import Client
-from zenml.constants import (
-    MODEL_METADATA_YAML_FILE_NAME,
-)
+from zenml.constants import MODEL_METADATA_YAML_FILE_NAME
 from zenml.enums import (
     ArtifactSaveType,
     ArtifactType,
@@ -45,10 +39,7 @@ from zenml.enums import (
     StackComponentType,
     VisualizationType,
 )
-from zenml.exceptions import (
-    DoesNotExistException,
-    StepContextError,
-)
+from zenml.exceptions import DoesNotExistException, StepContextError
 from zenml.io import fileio
 from zenml.logger import get_logger
 from zenml.metadata.metadata_types import validate_metadata
@@ -61,8 +52,6 @@ from zenml.models import (
     StepRunResponse,
     StepRunUpdate,
 )
-from zenml.stack import StackComponent
-from zenml.steps.step_context import get_step_context
 from zenml.utils import source_utils
 from zenml.utils.yaml_utils import read_yaml, write_yaml
 
@@ -147,6 +136,8 @@ def _store_artifact_data_and_prepare_request(
     Returns:
         Artifact version request for the artifact data that was stored.
     """
+    from zenml.client import Client
+
     artifact_store = Client().active_stack.artifact_store
     artifact_store.makedirs(uri)
 
@@ -235,6 +226,7 @@ def save_artifact(
     Returns:
         The saved artifact response.
     """
+    from zenml.client import Client
     from zenml.materializers.materializer_registry import (
         materializer_registry,
     )
@@ -320,9 +312,14 @@ def register_artifact(
         The saved artifact response.
 
     Raises:
-        FileNotFoundError: If the folder URI is outside of the artifact store
+        FileNotFoundError: If the folder URI is outside the artifact store
             bounds.
     """
+    from zenml.client import Client
+    from zenml.materializers.preexisting_data_materializer import (
+        PreexistingDataMaterializer,
+    )
+
     client = Client()
 
     # Get the current artifact store
@@ -381,6 +378,8 @@ def load_artifact(
     Returns:
         The loaded artifact.
     """
+    from zenml.client import Client
+
     artifact = Client().get_artifact_version(name_or_id, version)
     return load_artifact_from_response(artifact)
 
@@ -418,6 +417,7 @@ def log_artifact_metadata(
     )
 
     from zenml import log_metadata
+    from zenml.steps.step_context import get_step_context
 
     if artifact_name and artifact_version:
         assert artifact_name is not None
@@ -447,6 +447,8 @@ def log_artifact_metadata(
             infer_artifact=True,
         )
     elif artifact_name:
+        from zenml.client import Client
+
         client = Client()
         logger.warning(
             "Deprecation warning! Currently, you are calling "
@@ -628,6 +630,8 @@ def get_producer_step_of_artifact(
     Raises:
         RuntimeError: If the run that created the artifact no longer exists.
     """
+    from zenml.client import Client
+
     if not artifact.producer_step_run_id:
         raise RuntimeError(
             f"The run that produced the artifact with id '{artifact.id}' no "
@@ -678,6 +682,8 @@ def _check_if_artifact_with_given_uri_already_registered(
         RuntimeError: If the artifact store already contains an artifact with
             the given URI.
     """
+    from zenml.client import Client
+
     if artifact_store.exists(uri):
         # This check is only necessary for manual saves as we already check
         # it when creating the directory for step output artifacts
@@ -702,6 +708,9 @@ def _link_artifact_version_to_the_step_and_model(
     Args:
         artifact_version: The artifact version to link.
     """
+    from zenml.client import Client
+    from zenml.steps.step_context import get_step_context
+
     client = Client()
     try:
         error_message = "step run"
@@ -752,6 +761,8 @@ def _load_artifact_from_uri(
     Raises:
         ModuleNotFoundError: If the materializer or data type cannot be found.
     """
+
+    from zenml.client import Client
     from zenml.materializers.base_materializer import BaseMaterializer
 
     if not artifact_store:
@@ -818,6 +829,9 @@ def _load_artifact_store(
             an artifact store.
         NotImplementedError: If the artifact store could not be loaded.
     """
+    from zenml.client import Client
+    from zenml.stack import StackComponent
+
     if isinstance(artifact_store_id, str):
         artifact_store_id = UUID(artifact_store_id)
 
@@ -855,6 +869,9 @@ def _load_artifact_store(
 def _get_artifact_store_from_response_or_from_active_stack(
     artifact: ArtifactVersionResponse,
 ) -> "BaseArtifactStore":
+    from zenml.client import Client
+    from zenml.stack import StackComponent
+
     if artifact.artifact_store_id:
         try:
             artifact_store_model = Client().get_stack_component(
@@ -985,6 +1002,8 @@ def load_model_from_metadata(model_uri: str) -> Any:
     Returns:
         The ML model object loaded into memory.
     """
+    from zenml.client import Client
+
     # Load the model from its metadata
     artifact_versions_by_uri = Client().list_artifact_versions(uri=model_uri)
     if artifact_versions_by_uri.total == 1:
