+++ src/zenml/materializers/__init__.py
@@ -32,6 +32,9 @@ from zenml.materializers.structured_string_materializer import (
 from zenml.materializers.pydantic_materializer import PydanticMaterializer
 from zenml.materializers.service_materializer import ServiceMaterializer
 from zenml.materializers.uuid_materializer import UUIDMaterializer
+from zenml.materializers.preexisting_data_materializer import (
+    PreexistingDataMaterializer,
+)
 
 __all__ = [
     "BuiltInContainerMaterializer",
@@ -39,6 +42,7 @@ __all__ = [
     "BytesMaterializer",
     "CloudpickleMaterializer",
     "StructuredStringMaterializer",
+    "PreexistingDataMaterializer",
     "PydanticMaterializer",
     "ServiceMaterializer",
     "UUIDMaterializer",
